Fixes #1, no need to change method of storage
Believe that the actual cause of the issue was due to a bug when saving
the game to fen notation where any empty spots after the last piece were
not noted.  In any case, it has been tested to 160 moves without issues
and unless someone has a reproducible scenario, this issue is resolved.